# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IGeoLocation": {
				"properties": {
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"longitude",
					"latitude"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IAddress": {
				"properties": {
					"street": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					}
				},
				"required": [
					"street",
					"city",
					"postalCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RoomType": {
				"enum": [
					"bedroom",
					"bathroom",
					"garage"
				],
				"type": "string"
			},
			"IRoom": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/RoomType"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type",
					"count"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ConditionType": {
				"enum": [
					"perfect",
					"good",
					"bad"
				],
				"type": "string"
			},
			"IConstructionInfo": {
				"properties": {
					"type": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"condition": {
						"$ref": "#/components/schemas/ConditionType"
					}
				},
				"required": [
					"type",
					"year",
					"condition"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AmenityType": {
				"enum": [
					"wifi",
					"parking",
					"pool",
					"gym",
					"air_conditioning",
					"heating",
					"elevator",
					"washer",
					"dryer",
					"pet_friendly",
					"wheelchair_accessible",
					"security",
					"fireplace",
					"balcony",
					"garden",
					"sauna",
					"concierge",
					"barbecue_area",
					"storage",
					"smart_home"
				],
				"type": "string"
			},
			"IProperty": {
				"properties": {
					"geoLocation": {
						"$ref": "#/components/schemas/IGeoLocation"
					},
					"address": {
						"$ref": "#/components/schemas/IAddress"
					},
					"rooms": {
						"items": {
							"$ref": "#/components/schemas/IRoom"
						},
						"type": "array"
					},
					"constructionInfo": {
						"$ref": "#/components/schemas/IConstructionInfo"
					},
					"amenities": {
						"items": {
							"$ref": "#/components/schemas/AmenityType"
						},
						"type": "array"
					}
				},
				"required": [
					"geoLocation"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListingType": {
				"enum": [
					"sale",
					"rent"
				],
				"type": "string"
			},
			"IListing": {
				"properties": {
					"propertyData": {
						"$ref": "#/components/schemas/IProperty"
					},
					"type": {
						"$ref": "#/components/schemas/ListingType"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"propertyData",
					"type",
					"price"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IEmailVerifyPayload": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"IRegisterPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ILoginPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"rememberMe": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password",
					"rememberMe"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/listing/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IListing"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": []
			}
		},
		"/listing/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListing"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Put",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListing"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [
					{
						"jwt": [
							"broker"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IListing"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Listing"
				],
				"security": [
					{
						"jwt": [
							"broker"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/listing": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IListing"
								}
							}
						}
					}
				},
				"tags": [
					"Listing"
				],
				"security": [
					{
						"jwt": [
							"broker"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IListing"
							}
						}
					}
				}
			}
		},
		"/email/verify-email": {
			"put": {
				"operationId": "VerifyEmail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Email"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmailVerifyPayload"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginPayload"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/verify-email": {
			"post": {
				"operationId": "EmailVerify",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmailVerifyPayload"
							}
						}
					}
				}
			}
		},
		"/auth/profile": {
			"get": {
				"operationId": "Profile",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}